make it simpler
